Add node flow and pump group flow constraints.
Node flow constraint specifies that the sum of flows in/out of
the node needs to be equal zero in all calculated nodes.

Pump group flow constraint specifies that the flow in the elements
correponding to pump groups has to be equal to the sum of flows
of active pumps.