test: Add test for rolling out tensors into nx1 vectors
Checks the order (sequence) in which dimensions in a tensor are rolled
out into one big vector to ascertain that the A and b matrices/vectors
representing equality and inequality constraints are built properly.

Specifically, checks the order in which the loops have to be nested
to build A matrices out of rows and how the B matrices need to be
index before they can be 'rolled out' into vectors b.