fix: discrepancies between schedules from milp and simulator. Closes #6
1. Provides mapping between schedules (n and s) from milp programme
   outputs and what is supported by the simulator.
2. Plots individual pump speeds on the plot showing pump schedules
   from the mixed integer linear programme (the optimizer).
3. Splits schedule plotting function into two functions: one for the
   simulator and one for the optimizer.
4. Changes some plot formatting as some outputs were out of axis limits.

Clarification of 1.
The MILP formulation treats all pumps in a pump group individually whilst
the simulator treats pump groups with a single equation that assumes that
working pumps have the same pump speed. If the optimizer selects two or more
pumps ON, the speed that is fed back to the simulator is equal to the mean
of pump speeds of all pumps with status ON selected by the optimizer.